drivers/misc: mediatek: fpsgo_v3: Reset driver to a prior state
Mediatek pushed a bunch of commits to the driver after they dropped
mt6853 support from the tree. The newer commits are incompatible with
mt6853, which we brought back by reverting the removal commit and
therefore, it is proper to drop the incompatible changes to get
everything here working.

The driver was reset to bced957f880500be1dd714bbd10692cbe9089688

Signed-off-by: LinkBoi00 <linkdevel@protonmail.com>